requirements.txt
tensorflow==2.16.*
tf_keras>=2.16.0
numpy>=1.26.0,<2.0.0
scipy>=1.12.0
matplotlib>=3.8.3
tabulate>=0.9.0
nlpvocab>=1.2.0
absl-py>=1.4.0,<2.0.0
opencv-python>=4.9.0.80
